refactor(assertions): move assertions to test-helpers
